Shared config threshold logic. (#3693)
* Refactor to share threshold logic

* cleanup

* PR comments

* Update FormHelpers.jsx
